Trait circuit_construction::prologue::EvaluationDomain

pub trait EvaluationDomain<F>: Copy + Clone + Hash + Eq + PartialEq<Self> + Debug + CanonicalSerialize + CanonicalDeserializewhere
    F: FftField,{
    type Elements: Iterator<Item = F> + Sized;

    fn new(num_coeffs: usize) -> Option<Self>;
    fn compute_size_of_domain(num_coeffs: usize) -> Option<usize>;
    fn size(&self) -> usize;
    fn fft_in_place<T>(&self, coeffs: &mut Vec<T, Global>)
    where
        T: DomainCoeff<F>;
    fn ifft_in_place<T>(&self, evals: &mut Vec<T, Global>)
    where
        T: DomainCoeff<F>;
    fn evaluate_all_lagrange_coefficients(&self, tau: F) -> Vec<F, Global>;
    fn vanishing_polynomial(&self) -> SparsePolynomial<F>;
    fn evaluate_vanishing_polynomial(&self, tau: F) -> F;
    fn element(&self, i: usize) -> F;
    fn elements(&self) -> Self::Elements;

    fn sample_element_outside_domain<R>(&self, rng: &mut R) -> F
    where
        R: Rng,
    { ... }
    fn size_as_field_element(&self) -> F { ... }
    fn fft<T>(&self, coeffs: &[T]) -> Vec<T, Global>
    where
        T: DomainCoeff<F>,
    { ... }
    fn ifft<T>(&self, evals: &[T]) -> Vec<T, Global>
    where
        T: DomainCoeff<F>,
    { ... }
    fn distribute_powers<T>(coeffs: &mut [T], g: F)
    where
        T: DomainCoeff<F>,
    { ... }
    fn distribute_powers_and_mul_by_const<T>(coeffs: &mut [T], g: F, c: F)
    where
        T: DomainCoeff<F>,
    { ... }
    fn coset_fft<T>(&self, coeffs: &[T]) -> Vec<T, Global>
    where
        T: DomainCoeff<F>,
    { ... }
    fn coset_fft_in_place<T>(&self, coeffs: &mut Vec<T, Global>)
    where
        T: DomainCoeff<F>,
    { ... }
    fn coset_ifft<T>(&self, evals: &[T]) -> Vec<T, Global>
    where
        T: DomainCoeff<F>,
    { ... }
    fn coset_ifft_in_place<T>(&self, evals: &mut Vec<T, Global>)
    where
        T: DomainCoeff<F>,
    { ... }
    fn divide_by_vanishing_poly_on_coset_in_place(&self, evals: &mut [F]) { ... }
    fn reindex_by_subdomain(&self, other: Self, index: usize) -> usize { ... }
    fn mul_polynomials_in_evaluation_domain(
        &self,
        self_evals: &[F],
        other_evals: &[F]
    ) -> Vec<F, Global> { ... }

}

Defines a domain over which finite field (I)FFTs can be performed. The size of the supported FFT depends on the size of the multiplicative subgroup. For efficiency, we recommend that the field has at least one large subgroup generated by a root of unity.

Required Associated Types

type Elements: Iterator<Item = F> + Sized

The type of the elements iterator.

Required Methods

fn new(num_coeffs: usize) -> Option<Self>

Construct a domain that is large enough for evaluations of a polynomial having num_coeffs coefficients.

fn compute_size_of_domain(num_coeffs: usize) -> Option<usize>

Return the size of a domain that is large enough for evaluations of a polynomial having num_coeffs coefficients.

fn size(&self) -> usize

Return the size of self.

fn fft_in_place<T>(&self, coeffs: &mut Vec<T, Global>)where
T: DomainCoeff<F>,

Compute a FFT, modifying the vector in place.

fn ifft_in_place<T>(&self, evals: &mut Vec<T, Global>)where
T: DomainCoeff<F>,

Compute a IFFT, modifying the vector in place.

fn evaluate_all_lagrange_coefficients(&self, tau: F) -> Vec<F, Global>

Evaluate all the lagrange polynomials defined by this domain at the point tau. This is computed in time O(|domain|). Then given the evaluations of a degree d polynomial P over this domain, where d < |domain|, P(tau) can be computed as P(tau) = sum_{i in [|Domain|]} L_{i, Domain}(tau) * P(g^i). L_{i, Domain} is the value of the i-th lagrange coefficient in the returned vector.

fn vanishing_polynomial(&self) -> SparsePolynomial<F>

Return the sparse vanishing polynomial.

fn evaluate_vanishing_polynomial(&self, tau: F) -> F

This evaluates the vanishing polynomial for this domain at tau.

fn element(&self, i: usize) -> F

Returns the i-th element of the domain.

fn elements(&self) -> Self::Elements

Return an iterator over the elements of the domain.

Provided Methods

fn sample_element_outside_domain<R>(&self, rng: &mut R) -> Fwhere
R: Rng,

Sample an element that is not in the domain.

fn size_as_field_element(&self) -> F

Return the size of self as a field element.

fn fft<T>(&self, coeffs: &[T]) -> Vec<T, Global>where
T: DomainCoeff<F>,

Compute a FFT.

fn ifft<T>(&self, evals: &[T]) -> Vec<T, Global>where
T: DomainCoeff<F>,

Compute a IFFT.

fn distribute_powers<T>(coeffs: &mut [T], g: F)where
T: DomainCoeff<F>,

Multiply the i-th element of coeffs with g^i.

fn distribute_powers_and_mul_by_const<T>(coeffs: &mut [T], g: F, c: F)where
T: DomainCoeff<F>,

Multiply the i-th element of coeffs with c*g^i.

fn coset_fft<T>(&self, coeffs: &[T]) -> Vec<T, Global>where
T: DomainCoeff<F>,

Compute a FFT over a coset of the domain.

fn coset_fft_in_place<T>(&self, coeffs: &mut Vec<T, Global>)where
T: DomainCoeff<F>,

Compute a FFT over a coset of the domain, modifying the input vector in place.

fn coset_ifft<T>(&self, evals: &[T]) -> Vec<T, Global>where
T: DomainCoeff<F>,

Compute a IFFT over a coset of the domain.

fn coset_ifft_in_place<T>(&self, evals: &mut Vec<T, Global>)where
T: DomainCoeff<F>,

Compute a IFFT over a coset of the domain, modifying the input vector in place.

fn divide_by_vanishing_poly_on_coset_in_place(&self, evals: &mut [F])

The target polynomial is the zero polynomial in our evaluation domain, so we must perform division over a coset.

fn reindex_by_subdomain(&self, other: Self, index: usize) -> usize

Given an index which assumes the first elements of this domain are the elements of another (sub)domain, this returns the actual index into this domain.

fn mul_polynomials_in_evaluation_domain(
&self,
self_evals: &[F],
other_evals: &[F]
) -> Vec<F, Global>

Perform O(n) multiplication of two polynomials that are presented by their evaluations in the domain. Returns the evaluations of the product over the domain.

Assumes that the domain is large enough to allow for successful interpolation after multiplication.