Trait circuit_construction::prologue::PrimeField

pub trait PrimeField: Field<BasePrimeField = Self> + FftField<FftParams = Self::Params> + FromStr + From<Self::BigInt> + Into<Self::BigInt> + From<BigUint> + Into<BigUint> {
    type Params: FpParameters<BigInt = Self::BigInt>;
    type BigInt: BigInteger;

    fn from_repr(repr: Self::BigInt) -> Option<Self>;
    fn into_repr(&self) -> Self::BigInt;

    fn from_be_bytes_mod_order(bytes: &[u8]) -> Self { ... }
    fn from_le_bytes_mod_order(bytes: &[u8]) -> Self { ... }
    fn qnr_to_t() -> Self { ... }
    fn size_in_bits() -> usize { ... }
    fn trace() -> Self::BigInt { ... }
    fn trace_minus_one_div_two() -> Self::BigInt { ... }
    fn modulus_minus_one_div_two() -> Self::BigInt { ... }
}

The interface for a prime field.

Required Associated Types

type Params: FpParameters<BigInt = Self::BigInt>

type BigInt: BigInteger

Required Methods

fn from_repr(repr: Self::BigInt) -> Option<Self>

Returns a prime field element from its underlying representation.

fn into_repr(&self) -> Self::BigInt

Returns the underlying representation of the prime field element.

Provided Methods

fn from_be_bytes_mod_order(bytes: &[u8]) -> Self

Reads bytes in big-endian, and converts them to a field element. If the bytes are larger than the modulus, it will reduce them.

fn from_le_bytes_mod_order(bytes: &[u8]) -> Self

Reads bytes in little-endian, and converts them to a field element. If the bytes are larger than the modulus, it will reduce them.

fn qnr_to_t() -> Self

Return the QNR^t, for t defined by 2^s * t = MODULUS - 1, and t coprime to 2.

fn size_in_bits() -> usize

Returns the field size in bits.

fn trace() -> Self::BigInt

Returns the trace.

fn trace_minus_one_div_two() -> Self::BigInt

Returns the trace minus one divided by two.

fn modulus_minus_one_div_two() -> Self::BigInt

Returns the modulus minus one divided by two.