Trait circuit_construction::prologue::Sponge

pub trait Sponge<Input, Digest>where
    Input: Field,{
    // Required methods
    fn new(params: &'static ArithmeticSpongeParams<Input>) -> Self;
    fn absorb(&mut self, x: &[Input]);
    fn squeeze(&mut self) -> Digest;
    fn reset(&mut self);
}

Cryptographic sponge interface - for hashing an arbitrary amount of data into one or more field elements

Required Methods

fn new(params: &'static ArithmeticSpongeParams<Input>) -> Self

Create a new cryptographic sponge using arithmetic sponge params

fn absorb(&mut self, x: &[Input])

Absorb an array of field elements x

fn squeeze(&mut self) -> Digest

Squeeze an output from the sponge

fn reset(&mut self)

Reset the sponge back to its initial state (as if it were just created)