Function circuit_construction::prover::generate_prover_index

pub fn generate_prover_index<Curve, Circuit>(
    srs: Arc<SRS<Curve>>,
    public: usize,
    main: Circuit
) -> ProverIndex<Curve>where
    Circuit: FnOnce(&mut System<Curve::ScalarField>, Vec<Var<Curve::ScalarField>>),
    Curve: KimchiCurve,

Creates the prover index on input an srs, used constants, parameters for Poseidon, number of public inputs, and a specific circuit

Panics

Will panic if constraint_system is not built with public input.