Function circuit_construction::prover::prove

pub fn prove<G, H, EFqSponge, EFrSponge>(
    index: &ProverIndex<G>,
    group_map: &G::Map,
    blinders: Option<[Option<G::ScalarField>; 15]>,
    public_input: Vec<G::ScalarField>,
    main: H
) -> ProverProof<G> where
    H: FnMut(&mut WitnessGenerator<G::ScalarField>, Vec<Var<G::ScalarField>>),
    G::BaseField: PrimeField,
    G: KimchiCurve,
    EFqSponge: Clone + FqSponge<G::BaseField, G, G::ScalarField>,
    EFrSponge: FrSponge<G::ScalarField>, 

Given an index, a group map, custom blinders for the witness, a public input vector, and a circuit main, it creates a proof.