Trait circuit_construction::Cs

pub trait Cs<F: FftField + PrimeField> {

    fn var<G>(&mut self, g: G) -> Var<F>
    where
        G: FnOnce() -> F;
    fn curr_gate_count(&self) -> usize;
    fn gate(&mut self, g: GateSpec<F>);

    fn endo_scalar<G, N: BigInteger>(&mut self, length: usize, g: G) -> Var<F>
    where
        G: FnOnce() -> N,
    { ... }
    fn scalar<G, Fr: PrimeField>(
        &mut self, 
        length: usize, 
        g: G
    ) -> ShiftedScalar<F>
    where
        G: FnOnce() -> Fr,
    { ... }
    fn assert_eq(&mut self, x1: Var<F>, x2: Var<F>) { ... }
    fn constant(&mut self, x: F) -> Var<F> { ... }
    fn scale(&mut self, x: F, v: Var<F>) -> Var<F> { ... }
    fn add_group(
        &mut self, 
        zero: Var<F>, 
        (x1, y1): (Var<F>, Var<F>), 
        (x2, y2): (Var<F>, Var<F>)
    ) -> (Var<F>, Var<F>) { ... }
    fn double(
        &mut self, 
        zero: Var<F>, 
        (x1, y1): (Var<F>, Var<F>)
    ) -> (Var<F>, Var<F>) { ... }
    fn assert_add_group(
        &mut self, 
        zero: Var<F>, 
        (x1, y1): (Var<F>, Var<F>), 
        (x2, y2): (Var<F>, Var<F>), 
        (x3, y3): (Var<F>, Var<F>)
    ) { ... }
    fn cond_select(&mut self, b: Var<F>, t: Var<F>, f: Var<F>) -> Var<F> { ... }
    fn scalar_mul(
        &mut self, 
        zero: Var<F>, 
        (xt, yt): (Var<F>, Var<F>), 
        scalar: ShiftedScalar<F>
    ) -> (Var<F>, Var<F>) { ... }
    fn endo(
        &mut self, 
        zero: Var<F>, 
        constants: &Constants<F>, 
        (xt, yt): (Var<F>, Var<F>), 
        scalar: Var<F>, 
        length_in_bits: usize
    ) -> (Var<F>, Var<F>) { ... }
    fn assert_pack(&mut self, zero: Var<F>, x: Var<F>, bits_lsb: &[Var<F>]) { ... }
    fn poseidon(
        &mut self, 
        constants: &Constants<F>, 
        input: Vec<Var<F>>
    ) -> Vec<Var<F>> { ... }

}

Required Methods

fn var<G>(&mut self, g: G) -> Var<F> where
G: FnOnce() -> F,

In cases where you want to create a free variable in the circuit, as in the variable is not constrained yet and can be anything that the prover wants. For example, division can be implemented as:

let a = sys.constant(5u32.into());
let b = sys.constant(10u32.into());
let c = sys.var(|| {
   b.value * a.value.inverse().unwrap()
});
sys.assert_eq(a * c, b);

fn curr_gate_count(&self) -> usize

fn gate(&mut self, g: GateSpec<F>)

In circuit mode, adds a gate to the circuit. In witness generation mode, adds the corresponding row to the witness.

Provided Methods

fn endo_scalar<G, N: BigInteger>(&mut self, length: usize, g: G) -> Var<F> where
G: FnOnce() -> N,

fn scalar<G, Fr: PrimeField>(&mut self, length: usize, g: G) -> ShiftedScalar<F> where
G: FnOnce() -> Fr,

fn assert_eq(&mut self, x1: Var<F>, x2: Var<F>)

fn constant(&mut self, x: F) -> Var<F>

fn scale(&mut self, x: F, v: Var<F>) -> Var<F>

fn add_group(
&mut self,
zero: Var<F>,
(x1, y1): (Var<F>, Var<F>),
(x2, y2): (Var<F>, Var<F>)
) -> (Var<F>, Var<F>)

fn double(
&mut self,
zero: Var<F>,
(x1, y1): (Var<F>, Var<F>)
) -> (Var<F>, Var<F>)

fn assert_add_group(
&mut self,
zero: Var<F>,
(x1, y1): (Var<F>, Var<F>),
(x2, y2): (Var<F>, Var<F>),
(x3, y3): (Var<F>, Var<F>)
)

fn cond_select(&mut self, b: Var<F>, t: Var<F>, f: Var<F>) -> Var<F>

fn scalar_mul(
&mut self,
zero: Var<F>,
(xt, yt): (Var<F>, Var<F>),
scalar: ShiftedScalar<F>
) -> (Var<F>, Var<F>)

fn endo(
&mut self,
zero: Var<F>,
constants: &Constants<F>,
(xt, yt): (Var<F>, Var<F>),
scalar: Var<F>,
length_in_bits: usize
) -> (Var<F>, Var<F>)

fn assert_pack(&mut self, zero: Var<F>, x: Var<F>, bits_lsb: &[Var<F>])

fn poseidon(
&mut self,
constants: &Constants<F>,
input: Vec<Var<F>>
) -> Vec<Var<F>>