Struct circuit_construction::writer::WitnessGenerator

pub struct WitnessGenerator<F>where
    F: PrimeField,{
    pub generic_gate_queue: Vec<GateSpec<F>>,
    pub rows: Vec<[F; 15]>,
    pub cached_constants: HashMap<F, Var<F>>,
}

Carries a vector of rows corresponding to the witness, a queue of generic gates, and stores the cached constants

Fields
generic_gate_queue: Vec<GateSpec<F>>rows: Vec<[F; 15]>cached_constants: HashMap<F, Var<F>>
Implementations

impl<F> WitnessGenerator<F>where
F: PrimeField,

pub fn new(public_inputs: &[F]) -> Self

Given a list of public inputs, creates the witness generator.

impl<F: PrimeField> WitnessGenerator<F>

pub fn columns(&mut self) -> [Vec<F>; 15]

Returns the columns of the witness.