Trait circuit_construction::writer::Cs

pub trait Cs<F: PrimeField> {

    fn var<G>(&mut self, g: G) -> Var<F>
    where
        G: FnOnce() -> F;
    fn curr_gate_count(&self) -> usize;
    fn gate(&mut self, g: GateSpec<F>);
    fn cached_constants(&mut self, x: F) -> Var<F>;
    fn generic_queue(&mut self, gate: GateSpec<F>) -> Option<GateSpec<F>>;

    fn endo_scalar<G, N: BigInteger>(&mut self, length: usize, g: G) -> Var<F>
    where
        G: FnOnce() -> N,
    { ... }
    fn scalar<G, Fr: PrimeField>(
        &mut self,
        length: usize,
        g: G
    ) -> ShiftedScalar<F>
    where
        G: FnOnce() -> Fr,
    { ... }
    fn assert_eq(&mut self, x1: Var<F>, x2: Var<F>) { ... }
    fn constant(&mut self, x: F) -> Var<F> { ... }
    fn generic(&mut self, coeffs: [F; 5], vars: [Option<Var<F>>; 3]) { ... }
    fn scale(&mut self, x: F, v: Var<F>) -> Var<F> { ... }
    fn add_group(
        &mut self,
        zero: Var<F>,
        (x1, y1): (Var<F>, Var<F>),
        (x2, y2): (Var<F>, Var<F>)
    ) -> (Var<F>, Var<F>) { ... }
    fn double(
        &mut self,
        zero: Var<F>,
        (x1, y1): (Var<F>, Var<F>)
    ) -> (Var<F>, Var<F>) { ... }
    fn assert_add_group(
        &mut self,
        zero: Var<F>,
        (x1, y1): (Var<F>, Var<F>),
        (x2, y2): (Var<F>, Var<F>),
        (x3, y3): (Var<F>, Var<F>)
    ) { ... }
    fn cond_select(&mut self, b: Var<F>, t: Var<F>, f: Var<F>) -> Var<F> { ... }
    fn scalar_mul(
        &mut self,
        zero: Var<F>,
        (xt, yt): (Var<F>, Var<F>),
        scalar: ShiftedScalar<F>
    ) -> (Var<F>, Var<F>) { ... }
    fn endo(
        &mut self,
        zero: Var<F>,
        constants: &Constants<F>,
        (xt, yt): (Var<F>, Var<F>),
        scalar: Var<F>,
        length_in_bits: usize
    ) -> (Var<F>, Var<F>) { ... }
    fn assert_pack(&mut self, zero: Var<F>, x: Var<F>, bits_lsb: &[Var<F>]) { ... }
    fn poseidon(
        &mut self,
        constants: &Constants<F>,
        input: Vec<Var<F>>
    ) -> Vec<Var<F>> { ... }

}

This trait includes all the operations that can be executed by the elements in the circuits. It allows for different behaviours depending on the struct for which it is implemented for. In particular, the circuit mode and the witness generation mode.

Required Methods

fn var<G>(&mut self, g: G) -> Var<F> where
G: FnOnce() -> F,

In cases where you want to create a free variable in the circuit, as in the variable is not constrained yet and can be anything that the prover wants. For example, division can be implemented as:

let a = sys.constant(5u32.into());
let b = sys.constant(10u32.into());
let c = sys.var(|| {
   b.value * a.value.inverse().unwrap()
});
sys.assert_eq(a * c, b);

fn curr_gate_count(&self) -> usize

Returns the number of gates that the current Self contains.

fn gate(&mut self, g: GateSpec<F>)

In circuit mode, adds a gate to the circuit. In witness generation mode, adds the corresponding row to the witness.

fn cached_constants(&mut self, x: F) -> Var<F>

Checks if a constant x is already in the cached constants of self and returns it. Otherwise, it creates a variable for it and caches it.

fn generic_queue(&mut self, gate: GateSpec<F>) -> Option<GateSpec<F>>

Stores a generic gate until it can combine two of them into a double generic gate.

Provided Methods

fn endo_scalar<G, N: BigInteger>(&mut self, length: usize, g: G) -> Var<F> where
G: FnOnce() -> N,

Returns a variable containing a field element as value that is computed as the equivalent BigInteger number returned by function g, only if the length is a multiple of 4.

fn scalar<G, Fr: PrimeField>(&mut self, length: usize, g: G) -> ShiftedScalar<F> where
G: FnOnce() -> Fr,

This function creates a ShiftedScalar variable from a field element that is returned by function g(), and a length that should be a multiple of 5.

fn assert_eq(&mut self, x1: Var<F>, x2: Var<F>)

Creates a Generic gate that constrains if two variables are equal. This is done by setting x1 in the left wire and x2 in the right wire with left coefficient 1 and right coefficient -1, so that x1 - x2 = 0.

fn constant(&mut self, x: F) -> Var<F>

Creates a Generic gate to include a constant in the circuit, and returns the variable containing it. It sets the left wire to be the variable containing the constant x and the rest to zero. Then the left coefficient is set to one and the coefficient for constants is set to -x. This way, the constraint 1 * x - x = 0 holds.

fn generic(&mut self, coeffs: [F; 5], vars: [Option<Var<F>>; 3])

Adds a generic gate.

Warning: this assumes that some finalization occurs to flush any queued generic gate.

fn scale(&mut self, x: F, v: Var<F>) -> Var<F>

Creates a Generic gate to constrain that a variable v is scaled by an x amount and returns it. First, it creates a new variable with a scaled value (meaning, the value in v times x). Then, it creates a row that sets the left wire to be v and the right wire to be the scaled variable. Finally, it sets the left coefficient to x and the right coefficient to -1. That way, the constraint x * v - 1 * xv = 0 is created.

fn add_group(
&mut self,
zero: Var<F>,
(x1, y1): (Var<F>, Var<F>),
(x2, y2): (Var<F>, Var<F>)
) -> (Var<F>, Var<F>)

Performs curve point addition. It creates the corresponding CompleteAdd gate for the points (x1, y1) and (x2,y2) and returns the third point resulting from the addition as a tuple of variables.

fn double(&mut self, zero: Var<F>, (x1, y1): (Var<F>, Var<F>)) -> (Var<F>, Var<F>)

Doubles one curve point (x1, y1), using internally the add_group() function. It creates a CompleteAdd gate for this point addition (with itself). Returns a tuple of variables corresponding to the doubled point.

fn assert_add_group(
&mut self,
zero: Var<F>,
(x1, y1): (Var<F>, Var<F>),
(x2, y2): (Var<F>, Var<F>),
(x3, y3): (Var<F>, Var<F>)
)

Creates a CompleteAdd gate that checks whether a third point (x3, y3) is the addition of the two first points (x1, y1) and (x2, y2). The difference between this function and add_group() is that in assert_add_group the third point is given, whereas in the other one it is computed with the formula.

fn cond_select(&mut self, b: Var<F>, t: Var<F>, f: Var<F>) -> Var<F>

This function is used to include conditionals in circuits. It creates three Generic gates to simulate the logics of the conditional. It receives as input:

b: the branch

t: the true

f: the false And simulates the following equation: res = b * ( t - f ) + f ( when the condition is false, res = 1 ) ( when the condition is true, res = b ) This is constrained using three Generic gates

Constrain delta = t - f

Constrain res1 = b * delta

Constrain res = res1 + f For (1):

Creates a row with left wire t, right wire f, and output wire delta

Assigns 1 to the left coefficient, -1 to the right coefficient, and -1 to the output coefficient.

That way, it creates a first gate constraining: `1 * t - 1 * f - delta = 0`` For (2):

Creates a row with left wire b, right wire delta, and output wire res1.

Assigns -1 to the output coefficient, and 1 to the multiplication coefficient.

That way, it creates a second gate constraining: -1 * res + 1 * b * delta = 0 For (3):

Creates a row with left wire res1, right wire f, and output wire res.

Assigns 1 to the left coefficient, 1 to the right coefficient, and -1 to the output coefficient.

That way, it creates a third gate constraining: 1 * res1 + 1 * f - 1 * res = 0

fn scalar_mul(
&mut self,
zero: Var<F>,
(xt, yt): (Var<F>, Var<F>),
scalar: ShiftedScalar<F>
) -> (Var<F>, Var<F>)

Performs a scalar multiplication between a ShiftedScalar and a point (xt, yt). This function creates 51 rows pairs of rows.

fn endo(
&mut self,
zero: Var<F>,
constants: &Constants<F>,
(xt, yt): (Var<F>, Var<F>),
scalar: Var<F>,
length_in_bits: usize
) -> (Var<F>, Var<F>)

Creates an endoscalar multiplication gadget with length_in_bits/4 + 1 gates. For each row, it adds one EndoMul gate. The gadget is finalized with a Zero gate.

row	GateType

i	EndoMul

i+1	EndoMul

…	…

r	EndoMul

r+1	Zero

fn assert_pack(&mut self, zero: Var<F>, x: Var<F>, bits_lsb: &[Var<F>])

Checks that a string of bits (with LSB first) correspond to the value inside variable x. It splits the bitstring across rows, where each row takes care of 8 crumbs of 2 bits each.

fn poseidon(
&mut self,
constants: &Constants<F>,
input: Vec<Var<F>>
) -> Vec<Var<F>>

Creates a Poseidon gadget for given constants and a given input. It generates a number of Poseidon gates followed by a final Zero gate.