Module circuit_construction::prologue

This contains the Kimchi dependencies being used

Re-exports

pub use super::constants::fp_constants;

pub use super::constants::fq_constants;

pub use super::constants::Constants;

pub use super::prover::generate_prover_index;

pub use super::prover::prove;

pub use super::prover::CoordinateCurve;

pub use super::prover::FpInner;

pub use super::writer::Cs;

pub use super::writer::Var;

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

ArithmeticSponge

DefaultFqSponge

DefaultFrSponge

PlonkSpongeConstantsKimchi

PlonkSpongeConstantsLegacy

Radix2EvaluationDomain

Defines a domain over which finite field (I)FFTs can be performed. Works only for fields that have a large multiplicative subgroup of size that is a power-of-2.

SRS

VestaParameters

Traits

AffineCurve

Affine representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

CommitmentCurve

EvaluationDomain

Defines a domain over which finite field (I)FFTs can be performed. The size of the supported FFT depends on the size of the multiplicative subgroup. For efficiency, we recommend that the field has at least one large subgroup generated by a root of unity.

FftField

The interface for fields that are able to be used in FFTs.

GroupMap

PrimeField

The interface for a prime field.

ProjectiveCurve

Projective representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

Sponge

Cryptographic sponge interface - for hashing an arbitrary amount of data into one or more field elements

SpongeConstants

UniformRand

Functions

verify

Verify a proof ProverProof using a VerifierIndex and a group_map.

Type Definitions

Fp

PallasAffine

VestaAffine