Trait circuit_construction::prologue::ProjectiveCurve

pub trait ProjectiveCurve: 'static + Eq + ToBytes + FromBytes + CanonicalSerialize + CanonicalDeserialize + Copy + Clone + Default + Send + Sync + Hash + Debug + Display + UniformRand + Zeroize + Zero<Output = Self, Output = Self> + Neg<Output = Self> + Add<Self> + Sub<Self, Output = Self, Output = Self> + AddAssign<Self> + SubAssign<Self> + MulAssign<Self::ScalarField> + for<'a> Add<&'a Self> + for<'a> Sub<&'a Self> + for<'a> AddAssign<&'a Self> + for<'a> SubAssign<&'a Self> + Sum<Self> + for<'a> Sum<&'a Self> + From<Self::Affine> {
    type ScalarField: PrimeField + SquareRootField;
    type BaseField: Field;
    type Affine: AffineCurve + From<Self> + Into<Self>
    where
        <Self::Affine as AffineCurve>::Projective == Self,
        <Self::Affine as AffineCurve>::ScalarField == Self::ScalarField,
        <Self::Affine as AffineCurve>::BaseField == Self::BaseField;

    const COFACTOR: &'static [u64];

    fn prime_subgroup_generator() -> Self;
    fn batch_normalization(v: &mut [Self]);
    fn is_normalized(&self) -> bool;
    fn double_in_place(&mut self) -> &mut Self;
    fn add_assign_mixed(&mut self, other: &Self::Affine);

    fn batch_normalization_into_affine(v: &[Self]) -> Vec<Self::Affine, Global> { ... }
    fn double(&self) -> Self { ... }
    fn into_affine(&self) -> Self::Affine { ... }
    fn add_mixed(self, other: &Self::Affine) -> Self { ... }
    fn mul<S>(self, other: S) -> Self
    where
        S: AsRef<[u64]>,
    { ... }
}

Projective representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

Required Associated Types

type ScalarField: PrimeField + SquareRootField

type BaseField: Field

type Affine: AffineCurve + From<Self> + Into<Self>
where
<Self::Affine as AffineCurve>::Projective == Self,
<Self::Affine as AffineCurve>::ScalarField == Self::ScalarField,
<Self::Affine as AffineCurve>::BaseField == Self::BaseField

Required Associated Constants

const COFACTOR: &'static [u64]

Required Methods

fn prime_subgroup_generator() -> Self

Returns a fixed generator of unknown exponent.

fn batch_normalization(v: &mut [Self])

Normalizes a slice of projective elements so that conversion to affine is cheap.

fn is_normalized(&self) -> bool

Checks if the point is already “normalized” so that cheap affine conversion is possible.

fn double_in_place(&mut self) -> &mut Self

Doubles this element in place.

fn add_assign_mixed(&mut self, other: &Self::Affine)

Set self to be self + other, where other: Self::Affine. This is usually faster than adding other in projective form.

Provided Methods

fn batch_normalization_into_affine(v: &[Self]) -> Vec<Self::Affine, Global>

Normalizes a slice of projective elements and outputs a vector containing the affine equivalents.

fn double(&self) -> Self

Doubles this element.

fn into_affine(&self) -> Self::Affine

Converts self into the affine representation.

fn add_mixed(self, other: &Self::Affine) -> Self

Set self to be self + other, where other: Self::Affine. This is usually faster than adding other in projective form.

fn mul<S>(self, other: S) -> Self where
S: AsRef<[u64]>,

Performs scalar multiplication of this element.